+++ src/EarthModel/emfaultstickpainter.cc
@@ -245,7 +245,7 @@ bool FaultStickPainter::addPolyLine()
 				&& knotbinid.crl()==extrbid1.crl()) )
 			{
 			    const BinID bid = SI().transform( pos.coord() );
-			    const float z = zat ? zat->transform(pos) : pos.z;
+			    const double z = zat ? zat->transform(pos) : pos.z;
 			    if ( cs_.defaultDir() == CubeSampling::Inl )
 				stickauxdata->poly_ += FlatView::Point(
 								bid.crl(), z );
